+++ contracts/mocks/VoucherV2Mock.sol
@@ -4,12 +4,14 @@
 pragma solidity ^0.8.20;
 
 import {OwnableUpgradeable} from "@openzeppelin/contracts-upgradeable/access/OwnableUpgradeable.sol";
-import {IVoucher} from "../beacon/IVoucher.sol";
 
-contract VoucherV2Mock is OwnableUpgradeable, IVoucher {
+contract VoucherV2Mock is OwnableUpgradeable {
     /// @custom:storage-location erc7201:iexec.voucher.storage.Voucher
     struct VoucherStorage {
+        address _voucherHub;
         uint256 _expiration;
+        uint256 _type;
+        mapping(address => bool) _authorizedAccounts;
         uint256 _newStateVariable;
     }
 
@@ -27,12 +29,12 @@ contract VoucherV2Mock is OwnableUpgradeable, IVoucher {
      * Initialize new implementation contract.
      * @param newStateVariable test variable.
      */
-    function initialize(uint256 newStateVariable) external reinitializer(2) {
+    function initializeV2(uint256 newStateVariable) external reinitializer(2) {
         VoucherStorage storage $ = _getVoucherStorage();
         $._newStateVariable = newStateVariable;
     }
 
-    function getExpiration() external view override returns (uint256) {
+    function getExpiration() external view returns (uint256) {
         VoucherStorage storage $ = _getVoucherStorage();
         return $._expiration;
     }