+++ scripts/voucherHubUtils.ts
@@ -5,12 +5,22 @@ import { ContractFactory } from 'ethers';
 import { ethers, upgrades } from 'hardhat';
 import { VoucherHub } from '../typechain-types';
 
-export async function deployHub(iexecPoco: string, beacon: string): Promise<VoucherHub> {
+export async function deployHub(
+    assetEligibilityManager: string,
+    voucherManager: string,
+    iexecPoco: string,
+    beacon: string,
+): Promise<VoucherHub> {
     const VoucherHubFactory = await ethers.getContractFactory('VoucherHub');
     // @dev Type declaration produces a warning until feature is supported by
     // openzeppelin plugin. See "Support TypeChain in deployProxy function":
     // https://github.com/OpenZeppelin/openzeppelin-upgrades/pull/535
-    const contract: unknown = await upgrades.deployProxy(VoucherHubFactory, [iexecPoco, beacon]);
+    const contract: unknown = await upgrades.deployProxy(VoucherHubFactory, [
+        assetEligibilityManager,
+        voucherManager,
+        iexecPoco,
+        beacon,
+    ]);
     // Workaround openzeppelin-upgrades/pull/535;
     const voucherHub = contract as VoucherHub;
     return await voucherHub.waitForDeployment();